Unit2 / Unit 3

+++ APCS-A-2020/src/Unit2/Line.java
@@ -1,12 +1,13 @@
 package Unit2;
 //(c) A+ Computer Science
 //www.apluscompsci.com
-//Name -
-//Date -
+//Name - dylan g
+//Date - 2/4/20
+//Class - apcsa
 
 public class Line
 {
-	public static double getSlope( int x1, int y1, int x2, int y2 )
+	public static double getSlope( double x1, double y1, double x2, double y2 )
 	{
 		return ((y1 - y2) / (x1 - x2));
 	}
